feat(client): lands<->support width grip rides the cell edge (order-aware)
Replaces the floating ColumnSplitter (which inverted and overlapped the
draggable cells once reorder existed) with a ColumnEdgeHandle rendered on the
seam of the LEFT cell of the lands/support pair — only when they're adjacent.
It reads the live lands-col/support-col rects at drag time and flips the ratio
when support is the left cell, so reordering never inverts it. The old
overlay-measured columnBoundary + ColumnSplitter are removed.
